\name{Formula methods}
\alias{xyplot-methods}
\alias{hexbinplot-methods}
\alias{hexbinplot,formula,Raster-method}
\alias{xyplot,formula,Raster-method}
\title{Formula methods}
\description{Formula methods
}
\usage{
\S4method{xyplot}{formula,Raster}(x, data, dirXY, maxpixels=1e5,
            xlab='', ylab='', alpha=0.5, par.settings=rasterTheme,...)
\S4method{hexbinplot}{formula,Raster}(x, data, dirXY, maxpixels=1e5,
            xlab='', ylab='', par.settings=rasterTheme,...)

}
\arguments{
}
\value{
}
\details{
}
\references{
}
\seealso{
}
\examples{
}
\author{}
\keyword{}
