\name{rasterVis-package}
\alias{rasterVis-package}
\alias{rasterVis}
\docType{package}

\title{
  \packageTitle{rasterVis}
}


\description{
  \packageDescription{rasterVis}
}

\details{

The \code{rasterVis} package complements the \code{raster} and
\code{terra} packages, providing a set of methods for enhanced
visualization and interaction. This package defines visualization
methods with \code{levelplot} for quantitative data and categorical
data, both for univariate and multivariate rasters.

It also includes several methods in the frame of the Exploratory Data
Analysis approach: scatterplots with \code{xyplot}, histograms and
density plots with \code{histogram} and \code{densityplot}, violin and
boxplots with \code{bwplot}, and a matrix of scatterplots with
\code{splom}.

On the other hand, this package is able to display vector fields using
arrows, \code{vectorplot}, or with streamlines, \code{streamplot}.


\packageIndices{rasterVis}
}

\keyword{ package }
\keyword{spatial}

