\name{rasterVis-package}
\alias{rasterVis-package}
\alias{rasterVis}
\docType{package}
\title{Visualization methods for \code{raster}}
\description{The raster package defines classes and methods for spatial raster data access and manipulation. The rasterVis package complements raster providing a set of methods for enhanced visualization and interaction.
}
\details{
\tabular{ll}{
Package: \tab rasterVis\cr
Type: \tab Package\cr
Version: \tab 0.10\cr
Date: \tab 2011-06-17\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
Depends: \tab methods, raster, sp, grid, lattice, latticeExtra, zoo, hexbin, mgcv\cr
}
% ~~ An overview of how to use the package, including the most important ~~
% ~~ functions ~~
}
\author{
Oscar Perpi침치n Lamigueiro and Robert Hijmans

Maintainer: Oscar Perpi침치n Lamigueiro <oscar.perpinan@upm.es>
}
% \references{
% ~~ Literature or other references for background information ~~
% }
% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
% ~~ the R documentation directory ~~
\keyword{ package }
%\examples{}
